server/pyproject.toml
[project]
name = "clean-arch-template"
version = "1.0.0"
description = "A template for clean architecture projects"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.8",
]
